Changed search results table to fit on smaller screens
Changed the text size of search results so borders are not being cut off on screens smaller than 420px.

Aligned search results name test to the left, and evened out the spacing for other results.

Cut Search results name text to 10 characters before adding a break to continue on the next line.

Adjusted map border on dark mode to make up for the small text.

Changed map tiles, due to limit reach on that last tile.